// frontend/StockManager/generated-sources/openapi/models/TransactionDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * StockManager API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TransactionType } from './TransactionType';
import {
    TransactionTypeFromJSON,
    TransactionTypeFromJSONTyped,
    TransactionTypeToJSON,
    TransactionTypeToJSONTyped,
} from './TransactionType';
import type { StockDto } from './StockDto';
import {
    StockDtoFromJSON,
    StockDtoFromJSONTyped,
    StockDtoToJSON,
    StockDtoToJSONTyped,
} from './StockDto';

/**
 * 
 * @export
 * @interface TransactionDto
 */
export interface TransactionDto {
    /**
     * 
     * @type {number}
     * @memberof TransactionDto
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof TransactionDto
     */
    userId?: number;
    /**
     * 
     * @type {StockDto}
     * @memberof TransactionDto
     */
    stock?: StockDto;
    /**
     * 
     * @type {number}
     * @memberof TransactionDto
     */
    price?: number;
    /**
     * 
     * @type {Date}
     * @memberof TransactionDto
     */
    date?: Date;
    /**
     * 
     * @type {TransactionType}
     * @memberof TransactionDto
     */
    transactionType?: TransactionType;
    /**
     * 
     * @type {number}
     * @memberof TransactionDto
     */
    fee?: number;
    /**
     * 
     * @type {string}
     * @memberof TransactionDto
     */
    note?: string | null;
}



/**
 * Check if a given object implements the TransactionDto interface.
 */
export function instanceOfTransactionDto(value: object): value is TransactionDto {
    return true;
}

export function TransactionDtoFromJSON(json: any): TransactionDto {
    return TransactionDtoFromJSONTyped(json, false);
}

export function TransactionDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TransactionDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'userId': json['userId'] == null ? undefined : json['userId'],
        'stock': json['stock'] == null ? undefined : StockDtoFromJSON(json['stock']),
        'price': json['price'] == null ? undefined : json['price'],
        'date': json['date'] == null ? undefined : (new Date(json['date'])),
        'transactionType': json['transactionType'] == null ? undefined : TransactionTypeFromJSON(json['transactionType']),
        'fee': json['fee'] == null ? undefined : json['fee'],
        'note': json['note'] == null ? undefined : json['note'],
    };
}

export function TransactionDtoToJSON(json: any): TransactionDto {
    return TransactionDtoToJSONTyped(json, false);
}

export function TransactionDtoToJSONTyped(value?: TransactionDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'userId': value['userId'],
        'stock': StockDtoToJSON(value['stock']),
        'price': value['price'],
        'date': value['date'] == null ? undefined : ((value['date']).toISOString()),
        'transactionType': TransactionTypeToJSON(value['transactionType']),
        'fee': value['fee'],
        'note': value['note'],
    };
}